+++ server/v2/server_test.go
@@ -37,19 +37,9 @@ func (*mockApp[T]) InterfaceRegistry() coreapp.InterfaceRegistry {
 	return &mockInterfaceRegistry{}
 }
 
-// TODO split this test into multiple tests
-// test read config
-// test write config
-// test server configs
-// test start empty
-// test start config exists
-// test stop
 func TestServer(t *testing.T) {
 	currentDir, err := os.Getwd()
-	if err != nil {
-		t.Log(err)
-		t.Fail()
-	}
+	require.NoError(t, err)
 	configPath := filepath.Join(currentDir, "testdata")
 
 	v, err := serverv2.ReadConfig(configPath)
@@ -59,10 +49,8 @@ func TestServer(t *testing.T) {
 
 	logger := log.NewLogger(os.Stdout)
 	grpcServer := grpc.New[transaction.Tx]()
-	if err := grpcServer.Init(&mockApp[transaction.Tx]{}, v, logger); err != nil {
-		t.Log(err)
-		t.Fail()
-	}
+	err = grpcServer.Init(&mockApp[transaction.Tx]{}, v, logger)
+	require.NoError(t, err)
 
 	mockServer := &mockServer{name: "mock-server-1", ch: make(chan string, 100)}
 
@@ -73,30 +61,17 @@ func TestServer(t *testing.T) {
 	)
 
 	serverCfgs := server.Configs()
-	if serverCfgs[grpcServer.Name()].(*grpc.Config).Address != grpc.DefaultConfig().Address {
-		t.Logf("config is not equal: %v", serverCfgs[grpcServer.Name()])
-		t.Fail()
-	}
-	if serverCfgs[mockServer.Name()].(*mockServerConfig).MockFieldOne != MockServerDefaultConfig().MockFieldOne {
-		t.Logf("config is not equal: %v", serverCfgs[mockServer.Name()])
-		t.Fail()
-	}
+	require.Equal(t, serverCfgs[grpcServer.Name()].(*grpc.Config).Address, grpc.DefaultConfig().Address)
+	require.Equal(t, serverCfgs[mockServer.Name()].(*mockServerConfig).MockFieldOne, MockServerDefaultConfig().MockFieldOne)
 
 	// write config
-	if err := server.WriteConfig(configPath); err != nil {
-		t.Log(err)
-		t.Fail()
-	}
+	err = server.WriteConfig(configPath)
+	require.NoError(t, err)
 
 	v, err = serverv2.ReadConfig(configPath)
-	if err != nil {
-		t.Log(err) // config should be created by WriteConfig
-		t.FailNow()
-	}
-	if v.GetString(grpcServer.Name()+".address") != grpc.DefaultConfig().Address {
-		t.Logf("config is not equal: %v", v)
-		t.Fail()
-	}
+	require.NoError(t, err)
+
+	require.Equal(t, v.GetString(grpcServer.Name()+".address"), grpc.DefaultConfig().Address)
 
 	// start empty
 	ctx, cancelFn := context.WithCancel(context.TODO())
@@ -105,30 +80,10 @@ func TestServer(t *testing.T) {
 		<-time.After(5 * time.Second)
 		cancelFn()
 
-		if err := server.Stop(ctx); err != nil {
-			t.Logf("failed to stop servers: %s", err)
-			t.Fail()
-		}
+		err = server.Stop(ctx)
+		require.NoError(t, err)
 	}()
 
-	if err := server.Start(ctx); err != nil {
-		t.Log(err)
-		t.Fail()
-	}
-}
-
-func TestReadConfig(t *testing.T) {
-	currentDir, err := os.Getwd()
-	if err != nil {
-		t.Log(err)
-		t.Fail()
-	}
-	configPath := filepath.Join(currentDir, "testdata")
-
-	v, err := serverv2.ReadConfig(configPath)
-	require.NoError(t, err)
-
-	grpcConfig := grpc.DefaultConfig()
-	err = serverv2.UnmarshalSubConfig(v, "grpc", &grpcConfig)
+	err = server.Start(ctx)
 	require.NoError(t, err)
 }
